fbr_api_project/fbr_get_all_players/requirements.txt
mysql-connector-python==8.2.0
requests==2.31.0
pandas==2.1.0